requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.0.1
ibm-watson==8.1.0